-Added System.out.println() statements that are used for debugging purposes (in TestNode.java & Launcher0).
-In TestNode, added "value" field to be displayed in the received reply
string.

-Corrected the bug in TestNode of reading multiple bytes from the
inputstream (when the command issued is a GET).

-Implemented static methods in NodeCommands.java used to manipulate
byte[] to readable strings (requestByteArrayToString(byte[]) &
byteArrayToString(byte[]) )

--The socket inputstream in TestNode is now a BufferedInputStream
(instead of InputStream). (Originally thought InputStream might have
been the cause of the bugs, discovered that it was not but left the
modification to BufferedStream).

-Tests 0 to Test 8 have PASSED.

-Bug in the string representation used to display the ByteBuffer for the
field "Key" in the methods NodeCommands.get() & put(). Possible that
this bug affects the key used in the HashMap.